finished ssr and code splitting
installed new library, react imported component and react helmet
