+++ .babelrc
@@ -13,5 +13,16 @@
     "@babel/plugin-proposal-object-rest-spread",
     "@babel/plugin-transform-async-to-generator",
     "@babel/plugin-transform-runtime"
-  ]
+  ],
+  "env": {
+    "server": {
+      "plugins": [
+        "react-imported-component/babel",
+        "babel-plugin-dynamic-import-node"
+      ]
+    },
+    "client": {
+      "plugins": ["react-imported-component/babel"]
+    }
+  }
 }
