+++ package.json
@@ -41,6 +41,7 @@
     "@babel/register": "^7.14.5",
     "@parcel/reporter-bundle-buddy": "^2.0.0-nightly.1902",
     "@parcel/transformer-image": "^2.0.0-beta.2",
+    "babel-plugin-dynamic-import-node": "^2.3.3",
     "eslint": "^7.29.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.23.4",
@@ -57,7 +58,9 @@
     "mongoose": "^5.13.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-helmet": "^6.1.0",
     "react-image-gallery": "^1.2.4",
+    "react-imported-component": "^6.4.1",
     "react-loader-spinner": "^4.0.0",
     "react-router-dom": "^5.2.0",
     "react-scroll": "^1.8.2",
