+++ src/ui/html/index.html
@@ -30,6 +30,7 @@
   </head>
   <body>
     <div id="root"></div>
+
     <script src="../js/index.js"></script>
   </body>
 </html>
