--- src/ui/js/Homepage3.js
@@ -1,323 +1,0 @@
-import { Fragment, useRef } from "react";
-import { Link } from "react-router-dom";
-import { Link as ScrollLink } from "react-scroll";
-import { CSSTransition } from "react-transition-group";
-import { useScrollPosition } from "@n8tb1t/use-scroll-position";
-
-const Homepage3 = ({
-  toggleActive,
-  toggleAnimate3,
-  toggleAnimate2,
-  toggleNav,
-  animate,
-  halfHeight,
-  vnavState,
-  setActiveCache,
-}) => {
-  const page3 = useRef();
-  useScrollPosition(
-    ({ currPos }) => {
-      if (currPos.y <= halfHeight - 200 && currPos.y >= -halfHeight + 100) {
-        toggleActive(vnavState);
-        setActiveCache(vnavState);
-        toggleAnimate3(true);
-        toggleAnimate2(false);
-      }
-      if (currPos.y <= -1) {
-        toggleNav(false);
-      }
-      if (currPos.y > 0) {
-        toggleNav(true);
-      }
-    },
-    [],
-    page3
-  );
-
-  return (
-    <Fragment>
-      <div ref={page3} className="separator with-bg"></div>
-      <div className="page" id={vnavState}>
-        <svg
-          className="bg"
-          viewBox="0 0 1920 961"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <CSSTransition
-            in={animate}
-            timeout={400}
-            classNames="animate-left"
-            appear
-          >
-            <path
-              className="svg-bg-path"
-              d="M128.498 430C74.0676 430 31.5434 414.253 0.926067 382.76C-29.6913 351.267 -45 306.514 -45 248.5C-45 210.377 -38.3662 177.723 -25.0987 150.54C-11.491 123.025 7.38973 102.305 31.5434 88.3822C56.0373 74.1274 84.2734 67 116.252 67C135.643 67 155.204 72.4699 174.935 83.4096C182.419 87.3877 189.903 92.6918 197.388 99.3219C203.171 90.7027 209.124 83.9068 215.248 78.9342C221.371 74.6246 224.773 72.3041 225.453 71.9726H238.211V166.452H217.799C213.377 150.54 206.743 136.119 197.898 123.19C189.733 110.593 179.527 100.151 167.28 91.863C155.374 83.5753 141.596 79.4315 125.947 79.4315C102.474 79.4315 82.9126 92.6918 67.2637 119.212C51.955 145.733 44.3007 188.829 44.3007 248.5C44.3007 309.166 52.6354 352.593 69.3049 378.782C86.3145 404.64 109.448 417.568 138.704 417.568C154.693 417.568 169.832 413.259 184.12 404.64C198.748 395.689 210.485 383.921 219.33 369.334C228.515 354.748 233.108 339.333 233.108 323.089H245.865C245.865 341.985 240.592 359.721 230.046 376.296C219.84 392.54 205.722 405.634 187.692 415.579C170.002 425.193 150.271 430 128.498 430ZM396.839 430C377.788 430 360.609 426.188 345.3 418.563C330.331 410.607 318.595 400.33 310.09 387.733C301.585 375.136 297.333 361.875 297.333 347.952V188.829L266.715 181.37V173.911H333.053C363.67 173.911 378.979 188.829 378.979 218.664V347.952C378.979 368.837 382.721 384.915 390.205 396.186C398.03 407.126 407.045 412.596 417.251 412.596C429.838 412.596 441.234 405.966 451.44 392.705C455.182 387.733 459.095 381.103 463.177 372.815V188.829L432.56 181.37V173.911H575.441V181.37L544.823 188.829V425.027H539.72C514.886 425.027 496.005 418.397 483.078 405.137C477.295 399.17 473.213 392.54 470.831 385.247C466.068 393.534 460.455 400.993 453.992 407.623C438.683 422.541 419.632 430 396.839 430ZM606.257 415.082L636.874 407.623V188.829L606.257 181.37V173.911H652.183C672.255 173.911 688.074 180.21 699.64 192.807C704.063 197.779 707.805 203.912 710.867 211.205C715.969 203.912 722.263 197.282 729.747 191.315C746.757 178.055 765.808 171.425 786.9 171.425C804.93 171.425 818.367 175.568 827.213 183.856C836.398 191.812 840.99 203.415 840.99 218.664C840.99 229.936 837.418 239.052 830.274 246.014C823.47 252.644 814.285 255.959 802.719 255.959C797.956 255.959 793.193 255.627 788.43 254.964L782.307 253.473V183.856C764.957 183.856 749.138 190.486 734.85 203.747C728.046 210.377 722.603 217.007 718.521 223.637V407.623L749.138 415.082V422.541H606.257V415.082ZM863.897 415.082L894.514 407.623V188.829L863.897 181.37V173.911H909.823C929.894 173.911 945.713 180.21 957.28 192.807C961.702 197.779 965.444 203.912 968.506 211.205C973.609 203.912 979.902 197.282 987.387 191.315C1004.4 178.055 1023.45 171.425 1044.54 171.425C1062.57 171.425 1076.01 175.568 1084.85 183.856C1094.04 191.812 1098.63 203.415 1098.63 218.664C1098.63 229.936 1095.06 239.052 1087.91 246.014C1081.11 252.644 1071.92 255.959 1060.36 255.959C1055.6 255.959 1050.83 255.627 1046.07 254.964L1039.95 253.473V183.856C1022.6 183.856 1006.78 190.486 992.49 203.747C985.686 210.377 980.243 217.007 976.16 223.637V407.623L1006.78 415.082V422.541H863.897V415.082ZM1259.31 430C1215.76 430 1181.91 418.397 1157.76 395.192C1133.61 371.655 1121.53 339.333 1121.53 298.226C1121.53 271.374 1126.8 248.003 1137.35 228.112C1147.89 208.222 1162.52 192.973 1181.23 182.364C1200.28 171.756 1222.06 166.452 1246.55 166.452C1271.04 166.452 1292.48 170.927 1310.85 179.878C1329.56 188.829 1343.85 200.929 1353.71 216.178C1363.92 231.096 1369.02 247.671 1369.02 265.904C1369.02 274.192 1368.17 281.319 1366.47 287.286L1363.92 295.74C1349.29 298.723 1333.47 301.21 1316.46 303.199C1280.74 308.171 1244.68 310.658 1208.28 310.658C1208.28 349.112 1213.72 376.627 1224.61 393.203C1235.49 409.447 1250.46 417.568 1269.51 417.568C1282.78 417.568 1295.03 414.585 1306.25 408.618C1317.82 402.319 1326.84 394.197 1333.3 384.252C1340.1 373.975 1343.51 363.533 1343.51 352.925H1356.26C1356.26 366.185 1352.18 378.782 1344.02 390.716C1335.85 402.651 1324.45 412.264 1309.83 419.558C1295.2 426.519 1278.36 430 1259.31 430ZM1208.28 298.226C1230.05 298.226 1247.74 297.397 1261.35 295.74C1267.13 295.408 1273.26 294.579 1279.72 293.253C1280.06 291.596 1280.91 287.949 1282.27 282.314C1283.97 274.026 1284.82 265.241 1284.82 255.959C1284.82 228.444 1281.08 208.388 1273.6 195.79C1266.11 182.862 1257.1 176.397 1246.55 176.397C1221.04 176.397 1208.28 217.007 1208.28 298.226ZM1407.07 415.082L1437.69 407.623V188.829L1407.07 181.37V173.911H1453C1473.07 173.911 1488.89 180.21 1500.45 192.807C1504.88 197.779 1508.62 203.912 1511.68 211.205C1516.44 202.586 1522.06 195.127 1528.52 188.829C1543.83 173.911 1562.88 166.452 1585.67 166.452C1604.72 166.452 1621.73 170.43 1636.7 178.386C1652.01 186.011 1663.92 196.122 1672.42 208.719C1680.93 221.316 1685.18 234.577 1685.18 248.5V407.623L1715.8 415.082V422.541H1649.46C1618.84 422.541 1603.53 407.623 1603.53 377.788V248.5C1603.53 227.615 1599.62 211.703 1591.8 200.763C1584.31 189.492 1575.47 183.856 1565.26 183.856C1552.33 183.856 1540.77 190.486 1530.56 203.747C1526.14 209.382 1522.4 216.012 1519.34 223.637V407.623L1549.95 415.082V422.541H1407.07V415.082ZM1879.11 430C1856.66 430 1837.1 426.022 1820.43 418.066C1804.1 410.11 1791.51 399.17 1782.67 385.247C1773.82 371.323 1769.4 355.577 1769.4 338.007V186.342H1733.68V173.911H1769.4V149.048C1769.4 119.212 1784.71 104.295 1815.32 104.295H1851.05V173.911H1922.49V186.342H1851.05V338.007C1851.05 368.505 1854.28 389.39 1860.74 400.662C1867.2 411.933 1876.73 417.568 1889.32 417.568C1901.56 417.568 1912.28 411.767 1921.46 400.164C1930.65 388.562 1935.24 372.815 1935.24 352.925H1948C1948 377.456 1941.71 396.518 1929.12 410.11C1916.53 423.37 1899.86 430 1879.11 430Z"
-              fill="#484539"
-              fillOpacity="0.05"
-            />
-          </CSSTransition>
-          <CSSTransition
-            in={animate}
-            timeout={400}
-            classNames="animate-right"
-            appear
-          >
-            <path
-              className="svg-bg-path"
-              d="M1003.71 885C979.548 885 956.92 879.696 935.823 869.088C926.295 864.447 918.298 858.977 911.833 852.678C907.069 860.303 901.114 867.099 893.969 873.066C887.844 877.375 884.441 879.696 883.76 880.027H871V785.548H891.417C896.521 801.792 903.666 816.378 912.854 829.307C933.611 858.148 960.493 872.568 993.5 872.568C1012.9 872.568 1028.38 866.933 1039.95 855.662C1051.52 844.39 1057.3 829.307 1057.3 810.411C1057.3 799.471 1053.9 789.692 1047.09 781.073C1040.63 772.453 1032.46 764.829 1022.59 758.199C1012.73 751.569 998.944 743.281 981.25 733.336C958.791 721.07 940.587 709.964 926.635 700.019C912.684 690.074 900.774 677.643 890.906 662.725C881.038 647.475 876.104 629.574 876.104 609.021C876.104 592.777 880.698 578.19 889.885 565.262C899.413 552.001 913.535 541.559 932.25 533.934C950.965 525.978 973.934 522 1001.16 522C1020.89 522 1040.63 527.47 1060.36 538.41C1067.85 542.388 1075.34 547.692 1082.82 554.322C1088.61 545.703 1094.56 538.907 1100.69 533.934C1106.81 529.625 1110.21 527.304 1110.9 526.973H1123.66V621.452H1103.24C1098.82 605.54 1092.18 591.119 1083.33 578.19C1075.17 565.593 1064.96 555.151 1052.71 546.863C1040.8 538.575 1027.02 534.432 1011.36 534.432C998.093 534.432 986.694 539.57 977.166 549.847C967.639 560.123 962.875 573.218 962.875 589.13C962.875 601.396 966.448 612.501 973.593 622.447C981.079 632.06 990.097 640.514 1000.65 647.807C1011.53 655.1 1026.17 663.885 1044.54 674.162C1065.98 686.096 1083.33 696.704 1096.6 705.986C1109.87 715.269 1121.1 726.871 1130.29 740.795C1139.48 754.386 1144.07 770.133 1144.07 788.034C1144.07 805.936 1138.8 822.345 1128.25 837.263C1117.7 851.849 1101.88 863.452 1080.78 872.071C1059.68 880.69 1033.99 885 1003.71 885ZM1177.11 870.082L1207.73 862.623V544.377L1177.11 536.918V529.459H1243.46C1274.09 529.459 1289.4 544.377 1289.4 574.212V748.253L1383.83 651.288L1345.54 636.37V628.911H1442.52V636.37L1401.69 651.288L1342.99 710.959L1442.52 862.623L1468.04 870.082V877.541H1401.69C1385.36 877.541 1372.09 874.06 1361.88 867.099C1351.67 860.137 1341.97 849.529 1332.78 835.274L1289.4 768.144V862.623L1320.02 870.082V877.541H1177.11V870.082ZM1483.16 870.082L1513.78 862.623V643.829L1483.16 636.37V628.911H1549.51C1580.14 628.911 1595.45 643.829 1595.45 673.664V862.623L1626.07 870.082V877.541H1483.16V870.082ZM1549.51 599.075C1539.3 599.075 1530.8 595.76 1523.99 589.13C1517.19 582.5 1513.78 574.212 1513.78 564.267C1513.78 554.322 1517.19 546.034 1523.99 539.404C1530.8 532.774 1539.3 529.459 1549.51 529.459C1559.72 529.459 1568.23 532.774 1575.03 539.404C1581.84 546.034 1585.24 554.322 1585.24 564.267C1585.24 574.212 1581.84 582.5 1575.03 589.13C1568.23 595.76 1559.72 599.075 1549.51 599.075ZM1656.62 870.082L1687.25 862.623V544.377L1656.62 536.918V529.459H1722.97C1753.6 529.459 1768.91 544.377 1768.91 574.212V862.623L1799.54 870.082V877.541H1656.62V870.082ZM1830.08 870.082L1860.71 862.623V544.377L1830.08 536.918V529.459H1896.44C1927.06 529.459 1942.38 544.377 1942.38 574.212V862.623L1973 870.082V877.541H1830.08V870.082Z"
-              fill="#484539"
-              fillOpacity="0.05"
-            />
-          </CSSTransition>
-        </svg>
-
-        <div className="homepage-content">
-          <CSSTransition
-            in={animate}
-            timeout={400}
-            classNames="animate-left"
-            appear
-          >
-            <div className="text">
-              <h4>SKILLS</h4>
-              <h1>Current Skill Stack</h1>
-              <p>
-                Web development is a major, vast and growing <br /> expertise.
-                There was so much things happened since the <br /> beginning of
-                the web until what’s happening and <br /> emerging right now.
-              </p>
-              <p>
-                To caught up with all this technology, as a college student,{" "}
-                <br /> my approach is to keep learning and look up for all the{" "}
-                <br />
-                possibilities to growth and expand my experience with <br />{" "}
-                project, intership, and communities
-              </p>
-              <Link
-                to="/project"
-                onClick={() => {
-                  toggleActive("project");
-                }}
-                className="button-like"
-              >
-                See My Project
-              </Link>
-            </div>
-          </CSSTransition>
-          <CSSTransition
-            in={animate}
-            timeout={400}
-            classNames="animate-right"
-            appear
-          >
-            <svg
-              className="skill-stack"
-              viewBox="0 0 561 512"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                d="M455 459C455 429.729 478.729 406 508 406C537.271 406 561 429.729 561 459C561 488.271 537.271 512 508 512C478.729 512 455 488.271 455 459Z"
-                fill="#333333"
-              />
-              <path
-                d="M497.4 490.8C503.254 490.8 508 486.054 508 480.2V469.6H497.4C491.546 469.6 486.8 474.346 486.8 480.2C486.8 486.054 491.546 490.8 497.4 490.8Z"
-                fill="#0ACF83"
-              />
-              <path
-                d="M486.8 459C486.8 453.146 491.546 448.4 497.4 448.4H508V469.6H497.4C491.546 469.6 486.8 464.854 486.8 459Z"
-                fill="#A259FF"
-              />
-              <path
-                d="M486.8 437.8C486.8 431.946 491.546 427.2 497.4 427.2H508V448.4H497.4C491.546 448.4 486.8 443.654 486.8 437.8Z"
-                fill="#F24E1E"
-              />
-              <path
-                d="M508 427.2H518.6C524.454 427.2 529.2 431.946 529.2 437.8C529.2 443.654 524.454 448.4 518.6 448.4H508V427.2Z"
-                fill="#FF7262"
-              />
-              <path
-                d="M529.2 459C529.2 464.854 524.454 469.6 518.6 469.6C512.746 469.6 508 464.854 508 459C508 453.146 512.746 448.4 518.6 448.4C524.454 448.4 529.2 453.146 529.2 459Z"
-                fill="#1ABCFE"
-              />
-              <path
-                d="M284.941 442.893C278.601 442.893 275.451 446.202 275.451 450.967C275.451 454.035 276.985 456.208 279.015 457.202C278.021 458.114 277.193 458.912 276.447 459.698C275.7 460.487 275.244 461.319 275.244 462.232C275.244 463.351 275.7 464.389 277.068 464.967C275.618 466.337 274.788 467.33 274.788 469.196C274.788 472.47 277.647 474.377 285.148 474.377C292.36 474.377 296.544 471.685 296.544 466.918C296.544 461.408 292.318 461.2 285.148 460.743L281.087 460.495L282.455 458.299C283.2 458.506 284.403 458.671 285.731 458.671C290.951 458.671 294.557 455.978 294.557 451.377C294.557 449.471 293.615 448.145 292.87 447.191L296.531 446.901V443.234H288.754C287.802 443.234 286.723 442.893 284.941 442.893ZM286.806 466.504C288.796 466.628 290.371 466.711 290.371 468.039C290.371 470.274 287.096 470.442 284.694 470.442C281.792 470.442 279.679 469.944 279.679 468.287C279.679 467.457 280.217 466.753 280.757 466.091L286.806 466.504ZM285.192 454.32C282.455 454.32 281.337 452.91 281.337 450.88C281.337 448.187 283.035 447.067 285.315 447.067C287.677 447.067 289.086 448.352 289.086 450.797C289.086 452.953 287.511 454.32 285.192 454.32ZM304.127 439.661C306.531 439.661 307.899 438.293 307.899 435.929C307.899 433.569 306.531 432.283 304.127 432.283C301.765 432.283 300.397 433.569 300.397 435.929C300.397 438.292 301.765 439.661 304.127 439.661ZM307.297 460.174V443.234H299.237L298.781 446.461L300.99 447.184C301.405 447.308 301.5 447.535 301.5 447.99V460.174C301.5 460.796 301.15 460.921 300.57 461.004L298.187 461.294V464.766H310.609V461.294L308.227 461.004C307.604 460.921 307.297 460.839 307.297 460.174ZM325.862 459.312C324.866 459.685 323.682 459.974 322.52 459.974C320.283 459.974 319.719 459.022 319.719 456.576V446.547H325.515V442.406H320.018C319.686 442.406 319.719 442.404 319.719 442.114V436.609H315.749C315.294 440.75 313.094 442.569 309.781 442.984V446.547H313.096C313.428 446.547 313.094 446.839 313.094 447.087V457.944C313.094 462.629 315.126 464.617 320.056 464.617C322.42 464.617 325.114 463.83 326.897 462.96L325.862 459.312ZM246.627 433.136C245.536 432.045 243.765 432.045 242.674 433.136L238.739 437.071L243.731 442.062C244.891 441.67 246.221 441.933 247.146 442.858C248.075 443.788 248.336 445.129 247.934 446.293L252.745 451.104C253.909 450.703 255.251 450.962 256.181 451.893C257.481 453.192 257.481 455.297 256.181 456.596C254.881 457.896 252.777 457.896 251.477 456.596C250.5 455.618 250.132 454.183 250.628 452.979L246.015 448.493V460.299C246.015 460.456 246.757 460.665 247.021 460.928C248.319 462.227 248.382 464.332 247.083 465.633C245.784 466.931 243.709 466.931 242.411 465.633C241.112 464.332 241.128 462.227 242.427 460.928C242.748 460.608 242.703 460.365 243.531 460.203V448.287C242.703 448.124 242.763 447.883 242.443 447.561C241.459 446.577 241.222 445.134 241.727 443.926L236.806 439.004L223.813 451.997C222.721 453.089 222.721 454.86 223.813 455.95L242.763 474.899C243.853 475.989 245.623 475.989 246.715 474.899L265.576 456.037C266.666 454.946 266.666 453.175 265.576 452.084L246.627 433.136Z"
-                fill="#F34F29"
-              />
-              <path
-                d="M462.479 215.656L470.765 308.483L507.944 318.753L545.227 308.495L553.52 215.656H462.479ZM536.287 237.634L535.768 243.573L535.539 246.297H491.926L492.967 257.891H534.502L534.224 260.765L531.547 290.676L531.35 292.556L508 298.973V298.975L507.972 298.99L484.638 292.842L483.053 275.281H494.494L495.305 284.324L507.964 287.703H508V287.251L520.716 284.042L522.044 269.484H482.544L479.743 237.919L479.47 234.703H536.56L536.287 237.634Z"
-                fill="#E44D26"
-              />
-              <path
-                d="M62.4518 452.696C66.8364 452.696 70.3909 449.141 70.3909 444.756C70.3909 440.372 66.8364 436.817 62.4518 436.817C58.0672 436.817 54.5127 440.372 54.5127 444.756C54.5127 449.141 58.0672 452.696 62.4518 452.696Z"
-                fill="#61DAFB"
-              />
-              <path
-                d="M62.4516 473.952C68.6834 480.013 74.8298 483.598 79.7811 483.598C81.403 483.598 82.9396 483.257 84.2201 482.489C88.6592 479.928 90.2812 473.525 88.7446 464.391C88.4885 462.769 88.147 461.147 87.7202 459.44C89.4275 458.927 90.9641 458.415 92.5007 457.903C101.123 454.574 105.903 449.964 105.903 444.927C105.903 439.805 101.123 435.195 92.5007 431.951C90.9641 431.354 89.4275 430.842 87.7202 430.415C88.147 428.708 88.4885 427.086 88.7446 425.464C90.1958 416.159 88.5738 409.671 84.1348 407.11C82.8543 406.341 81.3177 406 79.6957 406C74.8298 406 68.598 409.585 62.3663 415.646C56.2199 409.585 50.0735 406 45.1222 406C43.5002 406 41.9636 406.341 40.6831 407.11C36.2441 409.671 34.6221 416.073 36.1587 425.207C36.4148 426.829 36.7563 428.451 37.1831 430.159C35.4758 430.671 33.9392 431.183 32.4026 431.695C23.7805 435.025 19 439.634 19 444.671C19 449.793 23.7805 454.403 32.4026 457.647C33.9392 458.244 35.4758 458.757 37.1831 459.183C36.7563 460.891 36.4148 462.513 36.1587 464.135C34.7075 473.269 36.3294 479.757 40.6831 482.232C41.9636 483.001 43.5002 483.342 45.1222 483.342C50.0735 483.513 56.2199 479.928 62.4516 473.952ZM57.6711 462.427C59.2077 462.513 60.8297 462.513 62.4516 462.513C64.0736 462.513 65.6956 462.513 67.2322 462.427C65.6956 464.476 64.0736 466.354 62.4516 468.147C60.8297 466.354 59.2077 464.476 57.6711 462.427ZM47.0856 453.635C47.9393 455.086 48.7076 456.452 49.6466 457.818C47.0003 457.476 44.5246 457.049 42.1344 456.537C42.9027 454.232 43.7563 451.842 44.7807 449.452C45.4637 450.817 46.232 452.269 47.0856 453.635ZM42.1344 433.061C44.5246 432.549 47.0003 432.122 49.6466 431.781C48.793 433.147 47.9393 434.512 47.0856 435.964C46.232 437.415 45.4637 438.781 44.7807 440.232C43.671 437.756 42.8173 435.366 42.1344 433.061ZM46.8295 444.756C47.9393 442.452 49.1344 440.147 50.5003 437.842C51.7808 435.622 53.232 433.403 54.6833 431.183C57.2443 431.012 59.8053 430.927 62.4516 430.927C65.1834 430.927 67.7444 431.012 70.22 431.183C71.7566 433.403 73.1225 435.622 74.403 437.842C75.7688 440.147 76.964 442.452 78.0737 444.756C76.964 447.061 75.7688 449.366 74.403 451.671C73.1225 453.891 71.6712 456.11 70.22 458.33C67.659 458.5 65.098 458.586 62.4516 458.586C59.7199 458.586 57.1589 458.5 54.6833 458.33C53.1467 456.11 51.7808 453.891 50.5003 451.671C49.1344 449.366 47.9393 447.061 46.8295 444.756ZM80.2079 440.147L77.903 435.878C77.0493 434.427 76.281 433.061 75.342 431.695C77.9884 432.037 80.464 432.464 82.8543 432.976C82.086 435.366 81.2323 437.756 80.2079 440.147ZM80.2079 449.366C81.2323 451.757 82.086 454.147 82.8543 456.452C80.464 456.964 77.9884 457.391 75.342 457.732C76.1957 456.366 77.0493 455 77.903 453.549C78.6713 452.269 79.4396 450.817 80.2079 449.366ZM82.1713 478.988C81.4884 479.415 80.6347 479.586 79.6957 479.586C75.5127 479.586 70.3054 476.171 65.1834 471.049C67.659 468.403 70.0493 465.415 72.4395 462.086C76.4518 461.744 80.2933 461.147 83.8787 460.293C84.3055 461.83 84.5616 463.366 84.8177 464.903C86.0128 472.159 85.0738 477.367 82.1713 478.988ZM86.6104 434C96.1715 436.732 101.891 440.915 101.891 444.756C101.891 448.086 97.9642 451.415 91.0495 454.061C89.6836 454.574 88.147 455.086 86.6104 455.513C85.5006 452.013 84.1348 448.427 82.4274 444.756C84.1348 441.086 85.586 437.5 86.6104 434ZM79.7811 409.927C80.7201 409.927 81.4884 410.098 82.2567 410.524C85.0738 412.146 86.0982 417.268 84.9031 424.61C84.647 426.061 84.3055 427.598 83.964 429.22C80.3786 428.451 76.5371 427.854 72.5249 427.427C70.22 424.098 67.7444 421.11 65.2687 418.464C70.3907 413.427 75.5981 409.927 79.7811 409.927ZM67.2322 427.086C65.6956 427 64.0736 427 62.4516 427C60.8297 427 59.2077 427 57.6711 427.086C59.2077 425.037 60.8297 423.159 62.4516 421.366C64.0736 423.159 65.6956 425.122 67.2322 427.086ZM42.7319 410.524C43.4149 410.098 44.2685 409.927 45.2076 409.927C49.3905 409.927 54.5979 413.342 59.7199 418.464C57.2443 421.11 54.854 424.098 52.4637 427.427C48.4515 427.768 44.61 428.366 41.0246 429.22C40.5978 427.683 40.3417 426.147 40.0856 424.61C38.8904 417.354 39.8295 412.232 42.7319 410.524ZM38.2929 455.513C28.7318 452.781 23.0122 448.598 23.0122 444.756C23.0122 441.427 26.9391 438.098 33.8538 435.451C35.2197 434.939 36.7563 434.427 38.2929 434C39.4026 437.5 40.7685 441.086 42.4758 444.756C40.7685 448.427 39.3173 452.098 38.2929 455.513ZM40.0856 464.903C40.3417 463.452 40.6831 461.915 41.0246 460.293C44.61 461.061 48.4515 461.659 52.4637 462.086C54.7686 465.415 57.2443 468.403 59.7199 471.049C54.5979 476.086 49.3905 479.586 45.2076 479.586C44.2685 479.586 43.5002 479.415 42.7319 478.988C39.8295 477.367 38.8904 472.159 40.0856 464.903ZM36.5002 500.074C38.3782 497.769 38.4636 495.208 37.4392 492.647C36.4148 490.086 34.2806 488.891 31.6343 488.806C28.4757 488.72 25.2318 488.806 22.0732 488.806H21.4756V510.916H24.0366V502.55H28.0489C28.5611 502.55 28.9879 502.72 29.244 503.147L34.366 511.086C34.4514 511.257 34.7075 511.513 34.8782 511.513H38.2075C36.1587 508.355 34.1953 505.367 32.1465 502.294C33.9392 502.037 35.4758 501.44 36.5002 500.074ZM24.0366 499.903V491.452H24.9757C26.9391 491.452 28.9879 491.367 30.9513 491.537C33.2562 491.623 34.8782 493.415 34.8782 495.72C34.8782 498.025 33.0001 499.818 30.6952 499.903C28.6464 499.989 26.5976 499.903 24.0366 499.903ZM57.0735 501.013C55.7077 495.037 50.2442 493.501 46.0612 495.379C42.8173 496.83 41.3661 499.647 41.2807 503.147C41.1953 505.794 41.9636 508.184 44.0124 509.977C46.3173 512.025 49.1344 512.281 52.0369 511.769C53.6589 511.428 55.1101 510.66 56.2199 509.464C55.793 508.867 55.3662 508.269 54.9394 507.757C52.5491 509.806 49.9027 510.489 47.0003 509.123C45.1222 508.184 44.1832 505.879 44.0124 504.172H57.2443V503.062C57.3296 502.294 57.2443 501.611 57.0735 501.013ZM44.1832 502.379C43.9271 499.818 46.4881 497.086 49.3052 497.086C52.5491 497.001 54.5979 498.964 54.6833 502.379H44.1832ZM70.3907 494.952C69.1102 494.696 67.7444 494.611 66.4639 494.696C64.4151 494.867 62.6224 495.806 61.1711 497.342C61.598 497.94 61.9394 498.537 62.4516 499.22C62.6224 499.05 62.7931 498.879 62.9638 498.794C64.3297 497.513 65.9517 496.83 67.9151 497.001C69.4517 497.086 70.9029 497.598 71.3298 499.135C71.6712 500.33 71.5859 501.611 71.6712 502.891C71.4151 502.891 71.3298 502.806 71.2444 502.72C69.1956 501.013 66.8907 500.672 64.4151 501.269C62.1102 501.867 60.6589 503.659 60.4882 505.964C60.3175 508.611 61.5126 510.574 63.8175 511.513C65.2687 512.111 66.8907 512.111 68.4273 511.769C69.6224 511.513 70.1346 510.83 71.842 509.891V511.001H74.2322C74.2322 507.586 74.3176 503.403 74.2322 499.476C74.2322 497.001 72.781 495.464 70.3907 494.952ZM71.5859 505.708C71.5005 505.964 71.5859 506.22 71.5859 506.477C71.5859 508.269 71.159 508.867 69.4517 509.55C68.2566 509.977 66.9761 510.147 65.6956 509.72C64.2443 509.294 63.2199 508.013 63.2199 506.562C63.1346 505.111 64.0736 503.659 65.5248 503.233C67.4883 502.55 69.281 502.977 70.9029 504.172C71.4151 504.598 71.7566 505.025 71.5859 505.708ZM84.9031 497.257C87.1226 496.574 89.1714 497.001 90.708 498.879L90.9641 499.05C91.5617 498.537 92.0739 498.025 92.7568 497.428C92.5007 497.172 92.4153 497.001 92.2446 496.745C89.769 494.098 84.9031 493.757 81.9152 495.891C77.7323 498.964 77.8176 504.94 79.8664 508.099C81.8299 511.086 84.647 512.111 87.9763 511.684C89.9397 511.428 91.5617 510.489 92.8422 508.867C92.2446 508.355 91.647 507.842 91.0495 507.245C90.8787 507.416 90.7934 507.501 90.708 507.672C88.4031 510.233 84.5616 509.977 82.5128 507.245C81.3177 505.623 81.0616 503.745 81.403 501.867C81.5738 499.733 82.6835 498.025 84.9031 497.257ZM102.659 508.867C102.489 508.952 102.403 509.038 102.403 509.038C101.72 509.55 101.037 509.635 100.269 509.379C99.5008 509.038 99.4154 508.355 99.33 507.672V497.94C99.33 497.769 99.33 498.111 99.4154 497.257H102.659V494.696H99.2447V490.428H96.6837V495.037H94.4641C94.2934 495.037 94.0373 495.208 94.0373 495.379C93.9519 495.976 94.0373 496.403 94.0373 497.257H96.769V508.184C96.769 509.55 97.1105 510.745 98.3056 511.428C99.5861 512.196 102.062 512.025 103.172 511.086C103.342 511.001 103.428 510.66 103.428 510.574C103.172 510.062 102.915 509.464 102.659 508.867Z"
-                fill="#61DAFB"
-              />
-              <path
-                d="M228.625 214L237.072 309.4L275 320L312.915 309.407L321.375 214H228.625Z"
-                fill="#2196F3"
-              />
-              <path
-                d="M303.315 245.197L302.249 257.182L299.029 293.109L275 299.754L274.98 299.76L250.965 293.109L249.289 274.32H261.055L261.929 284.066L274.987 287.59L274.994 287.584L288.071 284.059L289.853 267.821L248.725 267.934L247.559 256.83L290.841 256.32L291.715 244.435L246.4 244.561L245.638 233.504H304.362L303.315 245.197Z"
-                fill="#FAFAFA"
-              />
-              <path
-                d="M10.166 215.166H113.834V318.835H10.166V215.166Z"
-                fill="#F0DB4F"
-              />
-              <path
-                d="M105.35 294.11C104.591 289.38 101.507 285.408 92.3716 281.703C89.1982 280.245 85.6605 279.201 84.6063 276.796C84.232 275.396 84.1823 274.608 84.4191 273.761C85.099 271.011 88.3809 270.154 90.982 270.943C92.6573 271.504 94.2432 272.795 95.1996 274.855C99.6732 271.958 99.6641 271.977 102.788 269.986C101.644 268.213 101.034 267.395 100.285 266.636C97.5946 263.631 93.9285 262.083 88.0653 262.202L85.0112 262.597C82.0846 263.336 79.2955 264.873 77.66 266.932C72.7525 272.5 74.152 282.246 80.1236 286.256C86.0066 290.672 94.6489 291.677 95.7528 295.806C96.8269 300.86 92.0379 302.497 87.2778 301.915C83.7699 301.186 81.8188 299.403 79.7096 296.161C75.8273 298.408 75.8273 298.408 71.8358 300.703C72.7823 302.773 73.7769 303.709 75.3636 305.502C82.873 313.119 101.665 312.745 105.035 301.216C105.172 300.82 106.08 298.18 105.35 294.11ZM66.5233 262.812H56.8268L56.7871 287.881C56.7871 293.213 57.0628 298.1 56.1958 299.598C54.7772 302.545 51.1012 302.18 49.4259 301.608C47.7208 300.77 46.8537 299.578 45.8492 297.894C45.5734 297.41 45.3664 297.036 45.2968 297.007L37.4131 301.835C38.724 304.526 40.6552 306.861 43.1288 308.378C46.8239 310.596 51.7902 311.276 56.9842 310.083C60.3646 309.099 63.2812 307.059 64.8083 303.954C67.0161 299.884 66.5424 294.958 66.5225 289.508C66.5722 280.619 66.5233 271.73 66.5233 262.812Z"
-                fill="#323330"
-              />
-              <path
-                d="M547.423 92.2435C546.509 91.6428 545.564 91.0245 544.893 90.3532C544.504 89.9645 544.681 88.1581 544.778 87.1864C544.963 85.3226 545.149 83.4013 543.899 82.0189C543.74 81.8422 543.462 81.6744 543.205 81.5022C543.267 80.6012 543.334 79.7134 543.334 78.6667C543.334 64.3257 536.726 56.0533 531.422 49.4062C528.206 45.3783 525.667 42.1982 525.667 38.9167L525.671 37.8876C525.707 31.7926 525.777 19.1035 519.037 12.3239C516.184 9.4575 512.47 8 508 8C498.721 8 492.542 12.5094 492.542 34.5C492.542 40.5552 489.993 45.5107 487.299 50.7622C486.116 53.0632 484.897 55.4438 483.872 57.9613C482.021 62.5105 477.984 72.9736 477.242 79.0907C474.857 80.0933 474.164 82.7786 473.775 84.3774C473.643 84.9295 473.44 85.7554 473.418 85.8526C473.091 86.0557 471.854 86.2987 471.041 86.4621C468.343 86.9921 463.833 87.8842 463.833 91.9167C463.833 93.0915 464.01 94.0852 464.2 95.1232C464.54 96.9605 464.863 98.6962 463.904 102.406C463.856 102.587 463.833 102.773 463.833 102.958C463.833 108.267 472.079 108.956 477.008 109.367C477.83 109.438 478.51 109.486 478.947 109.557C479.627 109.663 481.05 110.321 482.428 110.952C485.369 112.313 489.035 114 492.542 114C494.154 114 495.514 113.276 496.543 112.083L496.592 112.123C497.572 111.032 504.078 109.583 508 109.583H521.52C521.763 110.436 522.116 111.284 522.792 112.026C523.984 113.337 525.693 114 527.875 114C533.608 114 536.386 111.045 539.076 108.192C539.531 107.711 539.994 107.216 540.48 106.73C541.598 105.613 543.788 104.813 545.723 104.107C548.89 102.95 552.167 101.753 552.167 98.5417C552.167 95.3572 549.539 93.6303 547.423 92.2435Z"
-                fill="#303C42"
-              />
-              <path
-                d="M487.962 59.622C488.911 57.2856 490.086 54.9934 491.226 52.7762C492.741 49.8259 494.264 46.8137 495.369 43.5631C498.054 45.9348 501.91 47.7501 503.584 47.7501C504.507 47.7501 511.86 45.2811 516.083 42.4545C524.563 59.8031 525.605 71.2335 525.658 81.9792C523.525 84.2891 522.421 89.8674 521.851 95.0128C515.354 101.563 502.365 101.121 496.654 100.211C495.545 98.1177 493.893 96.1214 491.597 94.5225C489.419 93.0076 487.732 89.4699 486.244 86.3517C484.804 83.344 483.501 80.6498 481.633 79.4043C482.247 74.8154 485.294 66.1808 487.962 59.622Z"
-                fill="white"
-              />
-              <path
-                d="M492.542 109.583C490.002 109.583 486.827 108.117 484.274 106.942C482.401 106.076 480.922 105.392 479.636 105.193C479.124 105.109 478.333 105.043 477.375 104.963C470.799 104.416 468.785 103.616 468.303 103.011C469.35 98.6696 468.904 96.2625 468.542 94.3192C468.4 93.5419 468.25 92.7954 468.228 91.9563C468.595 91.444 470.825 91.0023 471.903 90.7903C473.78 90.4193 475.401 90.0969 476.439 89.0589C477.419 88.0784 477.746 86.7269 478.068 85.4152C478.634 83.0832 478.868 83.0832 479.159 83.0611C479.941 83.41 481.332 86.325 482.255 88.2551C483.982 91.8679 485.934 95.9622 489.07 98.144C493.513 101.24 494.467 105.957 493.725 108.316C493.46 109.155 493.067 109.583 492.542 109.583ZM544.204 99.9549C541.841 100.816 539.164 101.792 537.358 103.603C536.837 104.124 536.346 104.645 535.86 105.162C533.445 107.728 531.7 109.579 527.875 109.579C526.987 109.579 526.378 109.398 526.055 109.049C525.393 108.32 525.552 106.593 525.623 105.856C525.649 105.56 525.667 105.321 525.667 105.167C525.684 102.274 526.329 94.3987 527.297 89.3372C527.473 89.5889 527.65 89.8451 527.866 90.0748C529.006 91.2805 530.534 91.9165 532.292 91.9165L533.069 91.9209C533.316 91.9209 533.568 91.9254 533.82 91.9254C536.028 91.9254 538.404 91.7399 540.29 89.9467C540.401 91.2187 540.772 92.4819 541.77 93.48C542.759 94.4694 543.952 95.2511 544.999 95.9401C546.659 97.031 547.746 97.7951 547.816 98.2809C547.467 98.7624 545.502 99.4823 544.204 99.9549ZM505.589 34.5043C509.577 35.3434 513.45 37.0792 514.404 38.3644C512.174 39.9589 505.081 43.1698 503.636 43.3332C502.113 43.1609 497.374 40.0781 496.936 39.0711C497.356 37.8079 502.245 34.6324 505.589 34.5043Z"
-                fill="#F68919"
-              />
-              <path
-                d="M513.724 32.6187C514.422 31.9562 514.908 30.8432 514.908 29.5137C514.908 27.4467 513.79 25.7728 512.412 25.7728C511.034 25.7728 509.917 27.4467 509.917 29.5137C509.917 30.1365 510.045 30.7018 510.226 31.2186C511.339 31.5807 512.549 32.0445 513.724 32.6187Z"
-                fill="#303C42"
-              />
-              <path
-                d="M512.416 22.0275C510.349 22.0275 508.676 24.5405 508.676 27.641C508.676 28.913 509.011 30.0305 509.484 30.9756C509.811 31.0772 510.323 31.2495 510.226 31.2185C510.045 30.7018 509.917 30.1409 509.917 29.5137C509.917 27.4467 511.034 25.7728 512.412 25.7728C513.79 25.7728 514.907 27.4467 514.907 29.5137C514.907 30.8431 514.417 31.9561 513.724 32.6186C513.817 32.6716 513.715 32.6186 513.949 32.7379C515.248 31.859 516.157 29.92 516.157 27.6455C516.157 24.5405 514.483 22.0275 512.416 22.0275ZM503.164 30.45C503.495 29.6109 503.738 28.6878 503.738 27.641C503.738 24.5405 502.064 22.0275 499.997 22.0275C497.93 22.0275 496.256 24.5405 496.256 27.641C496.256 29.7257 497.046 31.4703 498.173 32.4375C498.221 32.3757 498.332 32.3536 498.398 32.3227C497.864 31.6381 497.506 30.6532 497.506 29.5137C497.506 27.4467 498.623 25.7728 500.001 25.7728C501.379 25.7728 502.497 27.4467 502.497 29.5137C502.497 29.9289 502.422 30.3087 502.338 30.6797C502.532 30.6355 502.536 30.6002 503.164 30.45Z"
-                fill="white"
-              />
-              <path
-                d="M516.082 42.4545C511.86 45.2811 504.506 47.7501 503.583 47.7501C501.909 47.7501 498.053 45.9348 495.373 43.5675L525.309 73.5036C524.58 64.9927 522.394 55.3644 516.082 42.4545Z"
-                fill="url(#paint0_linear)"
-              />
-              <path
-                d="M547.423 92.2435C546.509 91.6428 545.564 91.0245 544.893 90.3532C544.504 89.9645 544.681 88.1581 544.778 87.1864C544.963 85.3226 545.149 83.4013 543.899 82.0189C543.74 81.8422 543.462 81.6744 543.205 81.5022C543.267 80.6012 543.334 79.7134 543.334 78.6667C543.334 64.3257 536.726 56.0533 531.422 49.4062C528.206 45.3783 525.667 42.1982 525.667 38.9167L525.671 37.8876C525.707 31.7926 525.777 19.1035 519.037 12.3239C516.184 9.4575 512.47 8 508 8C498.721 8 492.542 12.5094 492.542 34.5C492.542 40.5552 489.993 45.5107 487.299 50.7622C486.116 53.0632 484.897 55.4438 483.872 57.9613C482.021 62.5105 477.984 72.9736 477.242 79.0907C474.857 80.0933 474.164 82.7786 473.775 84.3774C473.643 84.9295 473.44 85.7554 473.418 85.8526C473.091 86.0557 471.854 86.2987 471.041 86.4621C468.343 86.9921 463.833 87.8842 463.833 91.9167C463.833 93.0915 464.01 94.0852 464.2 95.1232C464.54 96.9605 464.863 98.6962 463.904 102.406C463.856 102.587 463.833 102.773 463.833 102.958C463.833 108.267 472.079 108.956 477.008 109.367C477.83 109.438 478.51 109.486 478.947 109.557C479.627 109.663 481.05 110.321 482.428 110.952C485.369 112.313 489.035 114 492.542 114C494.154 114 495.514 113.276 496.543 112.083L496.592 112.123C497.572 111.032 504.078 109.583 508 109.583H521.52C521.763 110.436 522.116 111.284 522.792 112.026C523.984 113.337 525.693 114 527.875 114C533.608 114 536.386 111.045 539.076 108.192C539.531 107.711 539.994 107.216 540.48 106.73C541.598 105.613 543.788 104.813 545.723 104.107C548.89 102.95 552.167 101.753 552.167 98.5417C552.167 95.3572 549.539 93.6303 547.423 92.2435Z"
-                fill="url(#paint1_linear)"
-              />
-              <path
-                d="M316.675 69.8703C316.435 69.8703 316.197 69.8074 315.986 69.6848L313.792 68.3863C313.464 68.2033 313.624 68.1379 313.732 68.1006C314.169 67.9482 314.258 67.9126 314.724 67.6493C314.773 67.6219 314.837 67.6319 314.887 67.6617L316.573 68.6629C316.634 68.6969 316.721 68.6969 316.776 68.6629L323.349 64.8693C323.411 64.8345 323.45 64.7641 323.45 64.6912V57.1056C323.45 57.0311 323.411 56.9623 323.348 56.9242L316.778 53.1339C316.717 53.0991 316.636 53.0991 316.575 53.1339L310.006 56.9251C309.943 56.9623 309.902 57.0335 309.902 57.1056V64.6912C309.902 64.7641 309.943 64.8328 310.006 64.8668L311.806 65.9077C312.783 66.3955 313.382 65.8208 313.382 65.2427V57.7524C313.382 57.6472 313.467 57.5635 313.572 57.5635H314.406C314.51 57.5635 314.596 57.6472 314.596 57.7524V65.2427C314.596 66.5462 313.886 67.294 312.65 67.294C312.27 67.294 311.971 67.294 311.136 66.8833L309.413 65.8912C308.986 65.6444 308.724 65.1839 308.724 64.692V57.1064C308.724 56.6137 308.986 56.1541 309.413 55.9089L315.985 52.1103C316.401 51.876 316.954 51.876 317.366 52.1103L323.939 55.9089C324.364 56.1549 324.629 56.6145 324.629 57.1064V64.692C324.629 65.1848 324.364 65.6427 323.939 65.8912L317.366 69.6856C317.157 69.8074 316.92 69.8703 316.675 69.8703ZM318.707 64.644C315.829 64.644 315.227 63.3231 315.227 62.2159C315.227 62.1116 315.312 62.0263 315.417 62.0263H316.267C316.362 62.0263 316.441 62.0942 316.456 62.187C316.583 63.0524 316.967 63.4888 318.707 63.4888C320.09 63.4888 320.68 63.1757 320.68 62.4404C320.68 62.0164 320.513 61.7025 318.361 61.4913C316.563 61.3133 315.45 60.9166 315.45 59.4765C315.45 58.1507 316.568 57.359 318.442 57.359C320.548 57.359 321.589 58.0902 321.721 59.6587C321.726 59.7117 321.707 59.7639 321.671 59.8044C321.635 59.8417 321.585 59.8649 321.532 59.8649H320.678C320.59 59.8649 320.512 59.8028 320.494 59.7167C320.288 58.8057 319.792 58.5151 318.441 58.5151C316.93 58.5151 316.754 59.0426 316.754 59.4359C316.754 59.9138 316.961 60.0529 319.002 60.3229C321.022 60.5903 321.982 60.968 321.982 62.3882C321.983 63.8233 320.789 64.644 318.707 64.644Z"
-                fill="#83CD29"
-              />
-              <path
-                d="M303.141 59.6686C303.401 59.517 303.56 59.2404 303.56 58.9398V56.9904C303.56 56.6906 303.401 56.4141 303.141 56.2625L296.215 52.2411C295.954 52.0896 295.633 52.0888 295.371 52.2395L288.402 56.2633C288.142 56.4141 287.981 56.6915 287.981 56.9921V65.0381C287.981 65.3404 288.143 65.6203 288.407 65.7702L295.332 69.7162C295.587 69.8628 295.9 69.8636 296.158 69.7212L300.347 67.3925C300.48 67.318 300.563 67.1789 300.563 67.0281C300.564 66.8758 300.483 66.735 300.351 66.6596L293.339 62.6349C293.207 62.5596 293.127 62.4204 293.127 62.2706V59.7481C293.127 59.5974 293.207 59.4591 293.337 59.3845L295.521 58.1258C295.65 58.0504 295.811 58.0504 295.941 58.1258L298.125 59.3845C298.256 59.4599 298.336 59.5982 298.336 59.7481V61.7323C298.336 61.8838 298.416 62.0229 298.546 62.0975C298.677 62.1728 298.838 62.1728 298.969 62.0966L303.141 59.6686Z"
-                fill="#404137"
-              />
-              <path
-                d="M295.69 59.291C295.74 59.2629 295.802 59.2629 295.851 59.291L297.189 60.0637C297.238 60.0926 297.269 60.1465 297.269 60.2036V61.7481C297.269 61.806 297.238 61.859 297.189 61.888L295.851 62.6598C295.802 62.6888 295.739 62.6888 295.69 62.6598L294.353 61.888C294.303 61.859 294.272 61.806 294.272 61.7481V60.2036C294.272 60.1465 294.303 60.0926 294.353 60.0637L295.69 59.291Z"
-                fill="#83CD29"
-              />
-              <path
-                d="M277.553 62.5042C277.553 62.5787 277.513 62.6482 277.448 62.6863L275.054 64.066C274.989 64.1041 274.909 64.1041 274.844 64.066L272.451 62.6863C272.385 62.6482 272.346 62.5787 272.346 62.5042V59.7399C272.346 59.6653 272.386 59.595 272.451 59.5569L274.844 58.1755C274.91 58.1366 274.99 58.1366 275.055 58.1755L277.449 59.5569C277.514 59.595 277.554 59.6653 277.554 59.7399V62.5042H277.553ZM278.2 42.0362C278.07 41.9642 277.911 41.9658 277.781 42.0412C277.653 42.1174 277.574 42.2548 277.574 42.4039V53.0991C277.574 53.2035 277.517 53.3012 277.427 53.3542C277.336 53.4064 277.223 53.4064 277.132 53.3542L275.387 52.348C275.127 52.1981 274.806 52.1981 274.546 52.348L267.575 56.3711C267.314 56.5218 267.153 56.8 267.153 57.1006V65.1492C267.153 65.4498 267.314 65.7272 267.574 65.8788L274.545 69.9043C274.805 70.055 275.127 70.055 275.387 69.9043L282.36 65.8788C282.62 65.7272 282.781 65.4498 282.781 65.1492V45.0862C282.781 44.7815 282.615 44.4999 282.349 44.3508L278.2 42.0362Z"
-                fill="#404137"
-              />
-              <path
-                d="M253.666 52.1964C253.926 52.0457 254.247 52.0457 254.507 52.1964L261.478 56.2186C261.738 56.3693 261.899 56.6476 261.899 56.9482V64.9992C261.899 65.2998 261.738 65.5781 261.478 65.7288L254.508 69.7543C254.248 69.9042 253.927 69.9042 253.666 69.7543L246.697 65.7288C246.437 65.5781 246.276 65.2998 246.276 64.9992V56.9474C246.276 56.6476 246.437 56.3693 246.697 56.2186L253.666 52.1964Z"
-                fill="#83CD29"
-              />
-              <path
-                d="M240.989 56.8812C240.989 56.5781 240.83 56.2998 240.568 56.1499L233.6 52.1393C233.48 52.0698 233.349 52.0341 233.215 52.0292H233.143C233.011 52.0341 232.878 52.0698 232.759 52.1393L225.79 56.1499C225.53 56.2998 225.368 56.5781 225.368 56.8812L225.383 67.6799C225.383 67.8306 225.462 67.9706 225.593 68.0435C225.723 68.1213 225.882 68.1213 226.012 68.0435L230.152 65.6717C230.414 65.516 230.574 65.2419 230.574 64.9413V59.8964C230.574 59.5949 230.733 59.3175 230.994 59.1668L232.758 58.1515C232.888 58.0745 233.032 58.0381 233.178 58.0381C233.322 58.0381 233.47 58.0745 233.598 58.1515L235.36 59.1668C235.621 59.3175 235.782 59.5949 235.782 59.8964V64.9413C235.782 65.2419 235.943 65.5177 236.203 65.6717L240.344 68.0435C240.474 68.1213 240.636 68.1213 240.765 68.0435C240.893 67.9698 240.973 67.8298 240.973 67.6799L240.989 56.8812Z"
-                fill="#404137"
-              />
-              <path
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M80.803 31.23C78.199 23.122 74.022 15.946 68.136 9.771C66.648 8.209 64.994 6.778 63.956 4.835C63.3 3.605 62.675 2.358 62.036 1.12L61.63 0.098999L61.517 0.500999C61.464 2.521 60.32 3.89 58.896 5.169C57.292 6.607 55.8 8.169 54.26 9.678L49.524 15.907L45.695 22.949L43.134 29.864L43.057 29.971C41.648 34.6 40.953 39.36 40.612 44.166C40.483 45.973 40.631 47.805 40.732 49.621C40.877 52.217 41.328 54.768 42.004 57.281C44.461 66.407 49.448 73.976 56.267 80.408C57.533 81.603 58.902 82.69 60.223 83.826L60.808 85.834L61.352 88.95L61.612 92.203C61.609 92.863 61.582 93.526 61.621 94.184C61.632 94.353 61.852 94.509 61.976 94.671L63.08 95.059L64.229 95.506L64.032 92.615L64.023 89.767L64.42 85.429L64.708 84.485L65.533 83.024C66.551 82.206 67.642 81.462 68.569 80.551C70.246 78.904 71.92 77.234 73.421 75.429C75.368 73.088 77.044 70.533 78.39 67.793C79.289 65.96 80.137 64.09 80.838 62.175C81.456 60.487 81.839 58.712 82.326 56.975L82.454 56.6C83.459 51.912 83.628 47.176 83.259 42.41C82.962 38.569 82.059 34.862 80.803 31.23ZM60.903 81.505C61.057 80.734 61.248 79.967 61.387 79.193C61.248 79.967 61.058 80.734 60.903 81.505ZM64.32 82.037L63.674 80.622L64.32 82.037L65.269 82.848L64.32 82.037ZM104.474 99.964C103.648 98.381 102.436 97.179 100.834 96.39C99.492 95.73 98.049 95.44 96.565 95.398C95.453 95.366 94.337 95.423 93.223 95.437C92.234 95.449 91.244 95.466 90.255 95.457C89.092 95.445 87.929 95.41 86.766 95.377C86.573 95.371 86.436 95.41 86.346 95.606C86.205 95.911 86.038 96.205 85.865 96.539L86.059 96.601C86.636 96.703 87.216 96.79 87.79 96.905C88.528 97.052 88.86 97.476 88.894 98.098C88.944 98.984 88.964 99.872 88.961 100.76C88.946 104.274 88.921 107.788 88.895 111.301C88.893 111.533 88.889 111.775 88.826 111.993C88.753 112.245 88.674 112.571 88.486 112.695C88.161 112.911 87.758 113.057 87.371 113.12C86.81 113.212 86.716 113.237 86.541 113.789L86.465 114.065C86.381 114.366 86.426 114.425 86.74 114.428C88.542 114.448 90.343 114.487 92.144 114.481C93.787 114.475 95.43 114.575 97.067 114.266C98.614 113.975 100.058 113.465 101.376 112.602C103.086 111.481 104.316 109.983 104.965 108.028C105.489 106.449 105.606 104.838 105.428 103.187C105.307 102.048 105.001 100.975 104.474 99.964ZM101.317 109.625C100.353 111.419 98.915 112.617 96.86 112.933C95.573 113.13 94.284 113.073 93.057 112.586C92.28 112.278 91.991 111.607 91.967 110.814C91.944 110.037 91.935 109.259 91.934 108.482C91.93 105.748 91.93 103.014 91.934 100.281C91.936 99.42 91.951 98.557 91.965 97.695C91.975 97.089 92.102 96.886 92.693 96.837C95.289 96.619 97.766 96.899 99.823 98.726C101.095 99.856 101.819 101.297 102.12 102.952C102.245 103.642 102.283 104.348 102.361 105.048C102.298 106.646 102.082 108.201 101.317 109.625ZM123.106 106.664C122.594 105.418 121.624 104.637 120.405 104.137C119.989 103.966 119.56 103.825 119.111 103.659L119.268 103.559C119.753 103.248 120.293 102.997 120.711 102.614C121.727 101.683 122.149 100.512 121.951 99.121C121.763 97.798 121.103 96.827 119.924 96.197C118.854 95.627 117.7 95.419 116.506 95.42C114.44 95.422 112.373 95.453 110.307 95.457C109.595 95.458 108.883 95.405 108.171 95.395C108.033 95.393 107.828 95.428 107.769 95.52C107.606 95.77 107.498 96.058 107.382 96.336C107.315 96.498 107.381 96.587 107.566 96.611C108.063 96.679 108.559 96.764 109.057 96.838C109.745 96.941 110.078 97.299 110.12 97.992L110.129 98.403C110.13 100.558 110.137 102.713 110.128 104.868C110.117 107.019 110.093 109.172 110.067 111.324C110.064 111.595 110.027 111.867 109.988 112.136C109.929 112.542 109.712 112.822 109.296 112.91L108.119 113.142C107.98 113.17 107.779 113.166 107.722 113.252C107.506 113.575 107.332 113.928 107.356 114.354L107.477 114.387L111.43 114.484L112.223 114.481C113.591 114.465 114.961 114.492 116.327 114.422C117.806 114.348 119.195 113.909 120.479 113.154C121.846 112.349 122.898 111.288 123.272 109.692C123.512 108.673 123.513 107.648 123.106 106.664ZM113.063 97.483C113.069 97.05 113.26 96.862 113.69 96.851C114.749 96.822 115.801 96.828 116.823 97.193C118.145 97.665 118.958 98.805 118.943 100.198C118.936 100.733 118.944 101.263 118.747 101.777C118.358 102.789 117.612 103.323 116.554 103.427C116.002 103.483 115.445 103.489 114.953 103.515L113.311 103.443C113.093 103.435 112.998 103.339 112.999 103.115L113.063 97.483ZM119.869 110.977C119.34 112.128 118.376 112.733 117.169 112.943C116.204 113.111 115.231 113.062 114.277 112.816C113.571 112.635 113.283 112.068 113.142 111.439C113.047 111.018 113.063 110.517 113.055 110.079C113.042 109.403 113.052 108 113.052 108H113.038C113.038 107 113.041 106.134 113.035 105.175C113.034 104.968 113.069 104.865 113.322 104.873C114.22 104.9 115.121 104.915 116.019 104.95C116.822 104.981 117.574 105.219 118.281 105.6C119.357 106.18 120.005 107.068 120.183 108.288C120.319 109.213 120.261 110.123 119.869 110.977ZM21.282 112.055L21.301 106.618C21.304 105.8 21.2 104.998 20.932 104.222C20.193 102.085 18.155 101.112 16.033 101.879C15.068 102.228 14.203 102.757 13.377 103.357C12.896 103.707 12.896 103.707 12.548 103.208C11.563 101.796 10.156 101.313 8.51802 101.834C7.45902 102.17 6.53302 102.745 5.65602 103.413C5.57402 103.475 5.40902 103.544 5.36002 103.507C5.27802 103.446 5.22102 103.301 5.22302 103.192L5.28302 102.226C5.28802 102.023 5.24902 101.819 5.22902 101.606C4.83302 101.743 4.51702 101.88 4.18602 101.96C3.16302 102.21 2.13302 102.44 1.10402 102.675C0.855018 102.732 0.592018 102.807 0.568018 103.093C0.543018 103.374 0.814018 103.421 1.02402 103.505C1.46602 103.683 1.90502 103.872 2.34202 104.064C2.70902 104.226 2.84602 104.519 2.84402 104.913C2.83702 106.598 2.84802 108.281 2.83802 109.966C2.83402 110.651 2.80202 111.335 2.77102 112.02C2.74302 112.627 2.53602 112.881 1.94802 113.034C1.63602 113.116 1.31902 113.171 1.00502 113.245C0.936018 113.261 0.818018 113.305 0.817018 113.339C0.804018 113.636 0.788018 113.94 0.838018 114.229C0.848018 114.281 1.16202 114.281 1.33602 114.301L1.45302 114.294C2.66502 114.276 3.87702 114.257 5.09002 114.244C5.73302 114.237 6.37502 114.243 7.07302 114.243L7.14802 113.273C6.74802 113.2 6.39102 113.145 6.03902 113.068C5.49002 112.948 5.25602 112.657 5.24202 112.103L5.23202 111.31C5.22602 109.253 5.21802 107.197 5.21802 105.14C5.21802 104.841 5.34202 104.604 5.60502 104.425C6.16202 104.049 6.75002 103.75 7.40102 103.583C8.77302 103.232 9.96302 103.72 10.491 104.887C10.658 105.255 10.789 105.662 10.826 106.062C11.02 108.124 10.936 110.188 10.819 112.25C10.794 112.695 10.585 112.919 10.146 113.028L9.11402 113.246C9.03102 113.267 8.91002 113.211 8.90402 113.28C8.88102 113.565 8.89402 114.002 8.89402 114.002H9.14002L12.282 114.105C13.143 114.103 14.005 114.207 14.865 114.229C15.019 114.232 15.156 114.255 15.165 114.077C15.179 113.795 15.167 113.53 15.154 113.248L14.99 113.219L14.105 113.02C13.508 112.879 13.302 112.652 13.3 112.048C13.293 110.559 13.313 109.071 13.3 107.583C13.293 106.732 13.26 105.881 13.205 105.032C13.19 104.806 13.225 104.658 13.405 104.531C13.935 104.153 14.508 103.862 15.137 103.696C17.072 103.186 18.656 104.247 18.756 106.242C18.854 108.166 18.813 110.097 18.798 112.025C18.793 112.696 18.571 112.899 17.91 113.079L17.682 113.138C17.005 113.3 17.011 113.3 17.051 114.019C17.064 114.244 17.126 114.302 17.366 114.296C18.745 114.265 20.124 114.257 21.503 114.245C22.067 114.24 22.631 114.245 23.245 114.245L23.37 113.309C22.831 113.166 22.334 113.06 21.854 112.903C21.43 112.759 21.28 112.503 21.282 112.055ZM68.771 103.814C69.339 103.287 69.343 102.591 69.184 101.818C68.734 102.289 68.23 102.506 67.655 102.547C66.884 102.602 66.127 102.535 65.409 102.228C63.467 101.394 61.555 101.453 59.649 102.368C58.046 103.136 57.06 104.333 56.961 106.148C56.898 107.311 57.116 108.412 57.892 109.337C58.357 109.891 58.954 110.25 59.627 110.498C59.917 110.605 59.939 110.743 59.696 110.928C59.518 111.065 59.329 111.189 59.139 111.308C58.733 111.563 58.324 111.815 57.913 112.062C57.672 112.208 57.59 112.394 57.669 112.679C57.9 113.517 58.495 114.001 59.239 114.354L59.51 114.543L59.273 114.78C58.544 115.371 57.786 115.929 57.088 116.556C56.502 117.083 56.313 117.789 56.49 118.568C56.847 120.123 57.878 121.085 59.341 121.527C61.898 122.301 64.299 121.857 66.488 120.342C67.786 119.443 68.717 118.273 69 116.663C69.317 114.854 68.312 113.284 66.513 112.96C65.323 112.744 64.105 112.682 62.901 112.544C62.339 112.48 61.769 112.442 61.222 112.313C60.757 112.203 60.526 111.824 60.569 111.454C60.612 111.09 60.999 110.751 61.442 110.716C62.334 110.644 63.208 110.505 64.03 110.129C66.208 109.133 67.219 107.389 66.966 105.041C66.933 104.725 66.861 104.413 66.796 104.045C67.493 104.162 68.206 104.339 68.771 103.814ZM62.162 114.831C63.048 114.857 64.056 114.912 65.03 115.197C65.887 115.447 66.592 115.885 66.8 116.842C67.051 117.998 66.495 119.148 65.376 119.766C64.328 120.344 63.19 120.392 62.036 120.273C61.048 120.171 60.159 119.829 59.447 119.099C58.509 118.138 58.504 116.808 59.443 115.85C60.282 114.994 60.731 114.817 62.162 114.831ZM64.379 107.869C63.863 109.52 62.361 109.748 61.184 109.22C60.181 108.771 59.744 107.887 59.515 106.878C59.426 106.49 59.405 106.087 59.353 105.69C59.374 105.121 59.468 104.575 59.713 104.063C60.464 102.486 62.309 102.58 63.33 103.294C63.768 103.6 64.073 104.016 64.264 104.509C64.688 105.613 64.731 106.738 64.379 107.869ZM55.009 113.132C54.654 113.075 54.297 113.001 53.952 112.896C53.344 112.71 53.27 112.596 53.263 111.953C53.245 110.161 53.233 108.369 53.213 106.578C53.203 105.772 53.107 104.977 52.86 104.207C52.21 102.177 50.219 101.087 48.227 101.686C47.123 102.019 46.175 102.638 45.292 103.365L44.97 103.612L44.971 103.281C44.992 102.9 45.033 102.519 45.03 102.138C45.028 101.939 44.952 101.739 44.915 101.564C44.162 101.791 43.487 102.019 42.798 102.193C42.107 102.367 41.402 102.485 40.703 102.627C40.356 102.697 40.101 102.907 40.107 103.146C40.116 103.483 40.395 103.548 40.639 103.649C41.081 103.83 41.522 104.013 41.959 104.207C42.271 104.346 42.398 104.604 42.395 104.939C42.373 107.268 42.359 109.598 42.325 111.928C42.315 112.664 42.129 112.858 41.405 113.02L41.089 113.083C40.419 113.198 40.4 113.225 40.446 113.932L40.45 114.049C40.442 114.321 40.561 114.409 40.841 114.406C42.621 114.385 44.402 114.375 46.182 114.382C46.945 114.385 47.027 114.325 47.011 113.541L46.975 113.204C46.539 113.131 46.122 113.078 45.714 112.988C45.287 112.893 45.134 112.718 45.094 112.284C45.057 111.887 45.045 111.484 45.041 111.084C45.021 109.253 45.005 107.422 44.996 105.592C44.994 105.131 45.079 104.703 45.503 104.406C46.146 103.957 46.85 103.697 47.628 103.644C49.216 103.535 50.423 104.476 50.509 106.059C50.615 108.012 50.583 109.972 50.608 111.929C50.61 112.075 50.584 112.222 50.564 112.367C50.526 112.653 50.386 112.868 50.096 112.942C49.813 113.016 49.526 113.082 49.237 113.126C48.806 113.19 48.797 113.187 48.764 113.622L48.775 113.915C48.826 114.421 48.827 114.406 49.339 114.401C51.061 114.387 52.782 114.378 54.503 114.38C55.223 114.382 55.274 114.348 55.28 113.606L55.278 113.43C55.289 113.24 55.204 113.163 55.009 113.132ZM79.061 101.809C77.256 101.368 75.544 101.696 73.918 102.537C72.338 103.354 71.282 104.617 70.88 106.361C70.474 108.124 70.668 109.844 71.447 111.481C71.954 112.544 72.734 113.366 73.796 113.9C76.282 115.152 79.323 114.584 81.273 112.909C82.812 111.588 83.377 109.829 83.411 107.652C83.39 107.434 83.369 107.014 83.311 106.598C82.984 104.228 81.343 102.367 79.061 101.809ZM80.428 110.964C79.949 112.85 78.318 113.688 76.478 113.04C75.539 112.71 74.837 112.079 74.365 111.226C73.279 109.266 73.07 107.182 73.688 105.044C74.1 103.62 75.272 102.841 76.666 102.939C77.912 103.026 78.87 103.624 79.573 104.638C80.314 105.708 80.6 106.925 80.676 108.203C80.689 108.408 80.678 108.613 80.678 108.819L80.766 108.829C80.656 109.542 80.604 110.269 80.428 110.964ZM33.664 101.778C31.765 101.344 29.986 101.773 28.338 102.738C26.913 103.572 25.992 104.818 25.639 106.446C25.308 107.967 25.443 109.462 25.982 110.919C26.31 111.807 26.807 112.588 27.536 113.197C29.071 114.478 30.865 114.802 32.774 114.445C34.39 114.142 35.81 113.424 36.842 112.081C37.808 110.825 38.176 109.383 38.214 107.82C38.157 107.325 38.143 106.821 38.038 106.338C37.516 103.927 36.106 102.335 33.664 101.778ZM35.213 110.958C34.75 112.834 33.093 113.693 31.266 113.045C30.093 112.628 29.329 111.769 28.846 110.668C28.072 108.899 27.914 107.058 28.415 105.192C28.799 103.765 29.956 102.714 31.727 102.966C32.814 103.12 33.662 103.675 34.294 104.558C35.148 105.749 35.429 107.113 35.468 108.546V108.839L35.54 108.85C35.435 109.554 35.383 110.27 35.213 110.958Z"
-                fill="#4FAA41"
-              />
-              <line
-                x1="123"
-                y1="457.5"
-                x2="208"
-                y2="457.5"
-                stroke="#6B6755"
-                strokeWidth="3"
-              />
-              <line
-                x1="358"
-                y1="452.5"
-                x2="443"
-                y2="452.5"
-                stroke="#6B6755"
-                strokeWidth="3"
-              />
-              <line
-                x1="349"
-                y1="260.5"
-                x2="434"
-                y2="260.5"
-                stroke="#6B6755"
-                strokeWidth="3"
-              />
-              <line
-                x1="123"
-                y1="260.5"
-                x2="208"
-                y2="260.5"
-                stroke="#6B6755"
-                strokeWidth="3"
-              />
-              <line
-                x1="123"
-                y1="61.5"
-                x2="208"
-                y2="61.5"
-                stroke="#6B6755"
-                strokeWidth="3"
-              />
-              <line
-                x1="349"
-                y1="59.5"
-                x2="434"
-                y2="59.5"
-                stroke="#6B6755"
-                strokeWidth="3"
-              />
-              <defs>
-                <linearGradient
-                  id="paint0_linear"
-                  x1="500.827"
-                  y1="38.1085"
-                  x2="530.763"
-                  y2="68.049"
-                  gradientUnits="userSpaceOnUse"
-                >
-                  <stop stopOpacity="0.1" />
-                  <stop offset="1" stopOpacity="0" />
-                </linearGradient>
-                <linearGradient
-                  id="paint1_linear"
-                  x1="478.166"
-                  y1="52.4317"
-                  x2="556.963"
-                  y2="89.1739"
-                  gradientUnits="userSpaceOnUse"
-                >
-                  <stop stopColor="white" stopOpacity="0.2" />
-                  <stop offset="1" stopColor="white" stopOpacity="0" />
-                </linearGradient>
-              </defs>
-            </svg>
-          </CSSTransition>
-        </div>
-        <ScrollLink
-          to="intro"
-          onClick={() => {
-            toggleActive("intro");
-          }}
-          className="back-to-top button-like"
-        >
-          Back To Top
-        </ScrollLink>
-      </div>
-    </Fragment>
-  );
-};
-
-export default Homepage3;
